% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deck.R
\name{create_deck}
\alias{create_deck}
\title{Create the full quarto text for a phonics slidedeck in revealjs.}
\usage{
create_deck(let_1, let_2, let_3, title, image_dir, output_dir)
}
\arguments{
\item{let_1}{A vector of letters to be included as possible first letters.}

\item{let_2}{A vector of letters to be included as possible second letters.}

\item{let_3}{A vector of letters to be included as possible third letters.}

\item{title}{The title to be displayed as well as the title of the file.}

\item{image_dir}{The path to a folder where the letter images are stored.}

\item{output_dir}{The path to the folder where the quarto file will be saved.}
}
\value{
A long string with the text necessary for a complete slide in
reveal_js
}
\description{
Create the full quarto text for a phonics slidedeck in revealjs.
}
