% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slide.R
\name{create_slide}
\alias{create_slide}
\title{Create the quarto text for a phonics slide.}
\usage{
create_slide(word, image_dir)
}
\arguments{
\item{word}{A 3-letter nonsense word.}

\item{image_dir}{The path to a folder where the letter images are stored.}
}
\value{
A long string with the text necessary for a complete slide in
reveal_js
}
\description{
Create the quarto text for a phonics slide.
}
